fix: modified booking_participant.py (fixed autoincrement: Biginteger problem)

+++ server/app/models/booking_participant.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, Column, ForeignKey, Index, String
+from sqlalchemy import Integer, Column, ForeignKey, Index, String
 
 from app.db import Base
 
@@ -7,6 +7,6 @@ class BookingParticipant(Base):
     __tablename__ = "booking_participants"
     __table_args__ = (Index("ix_booking_participants_student", "student_id"),)
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True, nullable=False)
-    booking_id = Column(BigInteger, ForeignKey("bookings.id", ondelete="CASCADE"), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
+    booking_id = Column(Integer, ForeignKey("bookings.id", ondelete="CASCADE"), nullable=False)
     student_id = Column(String(20), ForeignKey("users.student_id", ondelete="CASCADE"), nullable=False)
